server/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings

# Supabase (se usar cliente python)
supabase>=1.0.0

# Para hashing de senha
passlib[bcrypt]

# Para manipulação de JWT (autenticação)
python-jose[cryptography]

# JWT (se usar python-jose ou pyjwt)
# python-jose[cryptography]
# pyjwt

# SQLAlchemy (se usar ORM completo)
# sqlalchemy
# psycopg2-binary
# alembic

# Para testes
# pytest
# httpx # Necessário para TestClient
